Refactoring sql script formatting helpers into To and From helpers
