+++ src/Microsoft.SqlTools.ServiceLayer/Metadata/SmoMetadataFactory.cs
@@ -10,7 +10,7 @@ using System.Data.SqlClient;
 using Microsoft.SqlServer.Management.Common;
 using Microsoft.SqlServer.Management.Smo;
 using Microsoft.SqlTools.ServiceLayer.Connection.ReliableConnection;
-using Microsoft.SqlTools.ServiceLayer.Utility;
+using Microsoft.SqlTools.ServiceLayer.Utility.SqlScriptFormatters;
 
 namespace Microsoft.SqlTools.ServiceLayer.Metadata
 {
@@ -92,13 +92,13 @@ namespace Microsoft.SqlTools.ServiceLayer.Metadata
                 // The default value may be escaped
                 string defaultValue = smoColumn.DefaultConstraint == null
                     ? null
-                    : SqlScriptFormatter.UnwrapLiteral(smoColumn.DefaultConstraint.Text);
+                    : FromSqlScript.UnwrapLiteral(smoColumn.DefaultConstraint.Text);
 
                 ColumnMetadata column = new ColumnMetadata
                 {
                     DefaultValue = defaultValue,
-                    EscapedName = SqlScriptFormatter.FormatIdentifier(smoColumn.Name),
-                    Ordinal = i,
+                    EscapedName = ToSqlScript.FormatIdentifier(smoColumn.Name),
+                    Ordinal = i
                 };
                 editColumns.Add(column);
             }
@@ -109,7 +109,7 @@ namespace Microsoft.SqlTools.ServiceLayer.Metadata
 
             // Escape the parts of the name
             string[] objectNameParts = {smoResult.Schema, smoResult.Name};
-            string escapedMultipartName = SqlScriptFormatter.FormatMultipartIdentifier(objectNameParts);
+            string escapedMultipartName = ToSqlScript.FormatMultipartIdentifier(objectNameParts);
 
             return new TableMetadata
             {
