+++ src/Microsoft.SqlTools.ServiceLayer/Utility/SqlScriptFormatters/Common.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.SqlTools.ServiceLayer.Utility.SqlScriptFormatters
+{
+    public class Common
+    {
+        
+    }
+}
